+++ app/javascript/pages/import.js
@@ -14,6 +14,13 @@ App.Pages.Import = (function () {
         data.append('csrf_token', vars('csrf_token'));
         data.append('file', $file[0].files[0]);
         data.append('import_type', $('#import-type').val());
+
+        const $imagesFile = $('#import-images-file');
+
+        if ($('#import-type').val() === 'ods' && $imagesFile[0].files.length) {
+            data.append('images_file', $imagesFile[0].files[0]);
+        }
+
         data.append('days_back', $('#days-back').val());
         data.append('days_forward', $('#days-forward').val());
         $('.import-phase:checked').each((index, el) => data.append('phases[]', $(el).val()));
@@ -152,8 +159,15 @@ App.Pages.Import = (function () {
             .fail(() => setExportWorking(false));
     }
 
+    function toggleImagesZip() {
+        $('#images-zip-wrapper').toggleClass('d-none', $('#import-type').val() !== 'ods');
+    }
+
     function initialize() {
         $('#export-data').on('click', onExportClick);
+
+        $('#import-type').on('change', toggleImagesZip);
+        toggleImagesZip();
 
         loadBackups();